pyproject.toml
[tool.poetry]
name = "test-paths"
version = "0.1.0"
description = ""
authors = ["oscar-cortez-home <oscarkortez@gmail.com>"]
readme = "README.md"
packages = [{include = "test_paths"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
